requirements.txt
streamlit
yagmail
pandas
python-dotenv
streamlit-quill
google-generativeai
fastapi
uvicorn
python-multipart
pypdf